% Generated by roxygen2 (4.0.2): do not edit by hand
\name{annotate.dataframe.clean}
\alias{annotate.dataframe.clean}
\title{You can semantically annotate data frames base on a tematres
thesaurus wih cleaning the dataset}
\usage{
annotate.dataframe.clean(input)
}
\arguments{
\item{input}{to be annotated takes a data frame}
}
\value{
a list wih annotation content
}
\description{
You can semantically annotate data frames base on a tematres
thesaurus wih cleaning the dataset
}

