\name{annotate.dataframe}
\alias{annotate.dataframe}
\title{This is the annotation features provided by rtematres}
\usage{
annotate.dataframe(input)
}
\arguments{
  \item{input}{to be annotated You can semantically
  annotate data frames base on a tematres thesaurus.}
}
\description{
This is the annotation features provided by rtematres
}

