\name{content.find.common.concept}
\alias{content.find.common.concept}
\title{This is the annotation features provided by rtematres
You can semantically annotate datasets base on a tematres
thesaurus.}
\usage{
content.find.common.concept(input, showbroader = F)
}
\description{
This is the annotation features provided by rtematres You
can semantically annotate datasets base on a tematres
thesaurus.
}

