\name{content.find.concept}
\alias{content.find.concept}
\title{This is the annotation features provided by rtematres}
\usage{
content.find.concept(input)
}
\arguments{
  \item{input}{to be annotated You can semantically
  annotate datasets base on a tematres thesaurus.}
}
\description{
This is the annotation features provided by rtematres
}

