\name{rtematres.api.define}
\alias{rtematres.api.define}
\title{This contains functions for user interaction with the api
More smart functions that wrap functionality of multiple
API tasks.}
\usage{
rtematres.api.define(term)
}
\arguments{
  \item{term}{It takes a term that you are searching the
  definition for}
}
\value{
Either a string or id
}
\description{
A tematres base api based function to retrieve definitions
}
\details{
It retrieves the definition of a term given the case it has
been described in the vocuabulary on the temates server.
}

