% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.api}
\alias{rtematres.api}
\title{Access to basic tematres server api}
\usage{
rtematres.api(task = "availableTasks", argument)
}
\arguments{
\item{task}{The api task you like to perform. Use the the task "availableTasks"
      to get an overview about the base api. It returns a data frame with
       descriptions and the arguments for the tasks.}

\item{argument}{Is the argument for the api task. You find the information
       about the type of arguments when you call the task "availableTasks".
       It depends on the task if the argument is numeric or a character.}
}
\value{
The function returns either a dataframe for "availableTasks" or a list
        of information elements for a certain task.
}
\description{
Features the tasks of the tematres server api. With no sugar added.
They are the basic building blocks for more convenient user functions.
}
\examples{
\dontrun{
    rtematres.api(task = "availableTasks")
    rtematres.api(task = "fetchVocabularyData")
    rtematres.api(task = "fetchTopTerms")
    rtematres.api(task = "search", argument = "measurement")
    rtematres.api(task = "fetch", argument = "measurement")
    rtematres.api(task = "searchNotes", argument = "measurement")
    rtematres.api(task = "suggest", argument = "measurement")
    rtematres.api(task = "suggestDetails", argument = "measurement")
    rtematres.api(task = "fetchSimilar", argument = "tre")
    rtematres.api(task = "letter", argument = "t")
    rtematres.api(task = "fetchTerm", argument = 12)
    rtematres.api(task = "fetchDown", argument = 4 )
    rtematres.api(task = "fetchUp", argument = 4)
    rtematres.api(task = "fetchRelated", argument = 4)
    rtematres.api(task = "fetchAlt", argument = 12 )
    rtematres.api(task = "fetchCode", argument = "tree")
    rtematres.api(task = "fetchNotes", argument = 5 )
    rtematres.api(task = "fetchDirectTerms", argument = 12)
    rtematres.api(task = "fetchURI", argument = 12)
    rtematres.api(task = "fetchTargetTerms", argument = 12 )
    rtematres.api(task = "fetchSourceTerm", argument = "measurement")
    rtematres.api(task = "fetchTerms", argument = '12,13' )
    rtematres.api(task = "fetchRelatedTerms", argument = '12,13' )
    rtematres.api(task = "fetchLast")
 }
}

