% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.common}
\alias{rtematres.common}
\title{Find common concept for categorial columns}
\usage{
rtematres.common(input)
}
\arguments{
\item{input}{A categorial vector}
}
\value{
The function returns the common concept
}
\description{
Search the thesaurus for concepts and extract their hirarchy.
Then return the common concept if any.
}
\examples{
\dontrun{
	rtematres.common(c("carbon", "nitrogen", "organic carbon"))
 }
}

