% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.define}
\alias{rtematres.define}
\title{Convenient functions for various tasks}
\usage{
rtematres.define(term)
}
\arguments{
\item{term}{The concept you are looking for}
}
\value{
It returns a text string describing the concept of interest
}
\description{
Define a concepts
}
\details{
The function retrieves the definition of a term. This works of course only given
the case it has been described in the vocuabulary you query.
}

