% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.hierarchy}
\alias{rtematres.hierarchy}
\title{Locate concepts in hierarchy}
\usage{
rtematres.hierarchy(term)
}
\arguments{
\item{term}{The concept you are looking for}
}
\value{
The function returns a character vector of concepts
}
\description{
Search the thesaurus for concepts and extract their hirarchy.
This envolves higher order and lower order terms.
}
\examples{
\dontrun{
	rtematres.hierarchy("carbon")
 }
}

