% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.illuminate}
\alias{rtematres.illuminate}
\title{Summarize information for a concept}
\usage{
rtematres.illuminate(input)
}
\arguments{
\item{input}{A concept name as string}
}
\value{
The function returns a list containing the information
}
\description{
Summarizes information for a concept in one convenient function
call. Currently it provides the definition, upstream and down
stream concepts for the term you querry for and related concepts
}
\examples{
\dontrun{
    rtematres.illuminate(input = "carbon")
 }
}

