% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.search}
\alias{rtematres.search}
\title{Search for concepts}
\usage{
rtematres.search(term)
}
\arguments{
\item{term}{The concept you are looking for}
}
\value{
The function returns a list of information for the results
}
\description{
Search the thesaurus for concepts. This function is a wrapper and so it
calls the appropriate funtions depending on the search task.
}

