% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rtematres.summary}
\alias{rtematres.summary}
\title{Create a summary for a vector}
\usage{
rtematres.summary(input)
}
\arguments{
\item{input}{A categorial vector}
}
\value{
The function returns a summary
}
\description{
Search the thesaurus for concepts in case of categorical textual vector.
Create a five value summary for numerical columns.
}
\examples{
\dontrun{
    rtematres.summary(input = iris$Species)
    lapply(iris, function(x) rtematres.summary(input = x))
 }
}

