\name{rtematres.api.define}
\alias{rtematres.api.define}
\title{A tematres base api based function to retrieve definitions}
\usage{
  rtematres.api.define(term)
}
\arguments{
  \item{term}{It takes a term that you are searching the
  definition for}
}
\value{
  Either a string or id
}
\description{
  It retrieves the definition of a term given the case it
  has been described in the vocuabulary on the temates
  server.
}

